distributor/requirements.txt
Django==1.9.5
django-crispy-forms==1.6.0
psycopg2==2.6.1
XlsxWriter==0.9.3
